requirements.txt
redis==5.2.0
Django==5.1.7
pydantic==2.10.5
python-dotenv==1.0.1
django-extensions==3.2.3
djangorestframework==3.15.2